# file_organizer_v2/docs/phase-6/plugin-development/api/openapi-plugin.json
{
  "components": {
    "schemas": {
...
      "FileInfo": {
        "properties": {
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "file_type": {
            "title": "File Type",
            "type": "string"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "modified": {
            "format": "date-time",
            "title": "Modified",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "path",
          "name",
          "size",
          "created",
          "modified",
          "file_type"
        ],
        "title": "FileInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HookEvent": {
        "description": "Canonical plugin hook events exposed via API + SDK.",
        "enum": [
          "file.scanned",
          "file.organized",
          "file.duplicated",
          "file.deleted",
          "organization.started",
          "organization.completed",
          "organization.failed",
          "deduplication.started",
          "deduplication.completed",
          "deduplication.found",
          "para.categorized",
          "johnny_decimal.assigned"
        ],
        "title": "HookEvent",
        "type": "string"
      },
...
      "PluginConfigValueResponse": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "key",
          "value"
        ],
        "title": "PluginConfigValueResponse",
        "type": "object"
      },
      "PluginFileListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FileInfo"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "PluginFileListResponse",
        "type": "object"
      },
      "PluginHookListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PluginHookRegistrationResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PluginHookListResponse",
        "type": "object"
      },
      "PluginHookRegistrationRequest": {
        "properties": {
          "callback_url": {
            "title": "Callback Url",
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/HookEvent"
          },
          "secret": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret"
          }
        },
        "required": [
          "event",
          "callback_url"
        ],
        "title": "PluginHookRegistrationRequest",
        "type": "object"
      },
      "PluginHookRegistrationResponse": {
        "properties": {
          "callback_url": {
            "title": "Callback Url",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/HookEvent"
          },
          "plugin_id": {
            "title": "Plugin Id",
            "type": "string"
          },
          "registered": {
            "title": "Registered",
            "type": "boolean"
          }
        },
        "required": [
          "plugin_id",
          "event",
          "callback_url",
          "created_at",
          "registered"
        ],
        "title": "PluginHookRegistrationResponse",
        "type": "object"
      },
      "PluginHookTriggerRequest": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/HookEvent"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "event"
        ],
        "title": "PluginHookTriggerRequest",
        "type": "object"
      },
      "PluginHookTriggerResponse": {
        "properties": {
          "delivered": {
            "title": "Delivered",
            "type": "integer"
          },
          "event": {
            "$ref": "#/components/schemas/HookEvent"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/PluginHookTriggerResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "event",
          "delivered",
          "failed",
          "results"
        ],
        "title": "PluginHookTriggerResponse",
        "type": "object"
      },
      "PluginHookTriggerResult": {
        "properties": {
          "callback_url": {
            "title": "Callback Url",
            "type": "string"
          },
          "delivered": {
            "title": "Delivered",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "event": {
            "$ref": "#/components/schemas/HookEvent"
          },
          "plugin_id": {
            "title": "Plugin Id",
            "type": "string"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          }
        },
        "required": [
          "plugin_id",
          "event",
          "callback_url",
          "status_code",
          "delivered"
        ],
        "title": "PluginHookTriggerResult",
        "type": "object"
      },
      "PluginHookUnregisterRequest": {
        "properties": {
          "callback_url": {
            "title": "Callback Url",
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/HookEvent"
          }
        },
        "required": [
          "event",
          "callback_url"
        ],
        "title": "PluginHookUnregisterRequest",
        "type": "object"
      },
      "PluginHookUnregisterResponse": {
        "properties": {
          "callback_url": {
            "title": "Callback Url",
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/HookEvent"
          },
          "plugin_id": {
            "title": "Plugin Id",
            "type": "string"
          },
          "removed": {
            "title": "Removed",
            "type": "boolean"
          }
        },
        "required": [
          "plugin_id",
          "event",
          "callback_url",
          "removed"
        ],
        "title": "PluginHookUnregisterResponse",
        "type": "object"
      },
      "PluginOrganizeFileRequest": {
        "properties": {
          "destination_path": {
            "title": "Destination Path",
            "type": "string"
          },
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "overwrite": {
            "default": false,
            "title": "Overwrite",
            "type": "boolean"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          }
        },
        "required": [
          "source_path",
          "destination_path"
        ],
        "title": "PluginOrganizeFileRequest",
        "type": "object"
      },
      "PluginOrganizeFileResponse": {
        "properties": {
          "destination_path": {
            "title": "Destination Path",
            "type": "string"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "moved": {
            "title": "Moved",
            "type": "boolean"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          }
        },
        "required": [
          "source_path",
          "destination_path",
          "moved",
          "dry_run"
        ],
        "title": "PluginOrganizeFileResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "REST API for AI-powered file organization",
    "title": "File Organizer API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/plugins/config/get": {
      "get": {
        "operationId": "get_config_for_plugins_api_v1_plugins_config_get_get",
        "parameters": [
          {
            "description": "Dot path to config value, e.g. updates.interval_hours",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "description": "Dot path to config value, e.g. updates.interval_hours",
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Profile",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginConfigValueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Config For Plugins",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/plugins/files/list": {
      "get": {
        "operationId": "list_files_for_plugins_api_v1_plugins_files_list_get",
        "parameters": [
          {
            "description": "Directory or file path",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "description": "Directory or file path",
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "recursive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Recursive",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_hidden",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Hidden",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "max_items",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Max Items",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginFileListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Files For Plugins",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/plugins/files/metadata": {
      "get": {
        "operationId": "get_file_metadata_for_plugins_api_v1_plugins_files_metadata_get",
        "parameters": [
          {
            "description": "File path",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "description": "File path",
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get File Metadata For Plugins",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/plugins/files/organize": {
      "post": {
        "operationId": "organize_file_for_plugins_api_v1_plugins_files_organize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginOrganizeFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginOrganizeFileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Organize File For Plugins",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/plugins/hooks": {
      "get": {
        "operationId": "list_plugin_hooks_api_v1_plugins_hooks_get",
        "parameters": [
          {
            "in": "query",
            "name": "event",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/HookEvent"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginHookListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Plugin Hooks",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/plugins/hooks/register": {
      "post": {
        "operationId": "register_plugin_hook_api_v1_plugins_hooks_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginHookRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginHookRegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Register Plugin Hook",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/plugins/hooks/trigger": {
      "post": {
        "operationId": "trigger_plugin_hook_event_api_v1_plugins_hooks_trigger_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginHookTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginHookTriggerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Trigger Plugin Hook Event",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/plugins/hooks/unregister": {
      "post": {
        "operationId": "unregister_plugin_hook_api_v1_plugins_hooks_unregister_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginHookUnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginHookUnregisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Unregister Plugin Hook",
        "tags": [
          "plugins"
        ]
      }
    }
  }
}